+++ crates/core/tests/plugin_test.rs
@@ -19,7 +19,7 @@ mod config {
     #[test]
     #[should_panic(expected = "InvalidPluginLocator")]
     fn errors_for_empty_location() {
-        PluginLocator::from_str("schema:").unwrap();
+        PluginLocator::from_str("source:").unwrap();
     }
 
     mod schema {
@@ -28,53 +28,41 @@ mod config {
         #[test]
         fn deser_to_enum() {
             let value: PluginLocator =
-                serde_json::from_str("\"schema:https://foo.com/file.toml\"").unwrap();
+                serde_json::from_str("\"source:https://foo.com/file.toml\"").unwrap();
 
             assert_eq!(
                 value,
-                PluginLocator::Schema(PluginLocation::Url("https://foo.com/file.toml".into()))
+                PluginLocator::Source(PluginLocation::Url("https://foo.com/file.toml".into()))
             );
         }
 
         #[test]
         fn ser_to_string() {
             assert_eq!(
-                serde_json::to_string(&PluginLocator::Schema(PluginLocation::Url(
+                serde_json::to_string(&PluginLocator::Source(PluginLocation::Url(
                     "https://foo.com/file.toml".into()
                 )))
                 .unwrap(),
-                "\"schema:https://foo.com/file.toml\""
+                "\"source:https://foo.com/file.toml\""
             );
         }
 
         #[test]
         #[should_panic(expected = "InvalidPluginLocator")]
         fn errors_for_http_url() {
-            PluginLocator::from_str("schema:http://foo.com").unwrap();
+            PluginLocator::from_str("source:http://foo.com").unwrap();
         }
 
         #[test]
         #[should_panic(expected = "InvalidPluginLocator")]
         fn errors_for_abs_path() {
-            PluginLocator::from_str("schema:/foo/file").unwrap();
+            PluginLocator::from_str("source:/foo/file").unwrap();
         }
 
         #[test]
         #[should_panic(expected = "InvalidPluginLocator")]
         fn errors_for_random_value() {
-            PluginLocator::from_str("schema:abc123").unwrap();
-        }
-
-        #[test]
-        #[should_panic(expected = "InvalidPluginLocatorExt(\".toml\")")]
-        fn errors_for_invalid_url_ext() {
-            PluginLocator::from_str("schema:https://foo.com/file.yaml").unwrap();
-        }
-
-        #[test]
-        #[should_panic(expected = "InvalidPluginLocatorExt(\".toml\")")]
-        fn errors_for_invalid_path_ext() {
-            PluginLocator::from_str("schema:../foo/file.yaml").unwrap();
+            PluginLocator::from_str("source:abc123").unwrap();
         }
     }
 
@@ -122,13 +110,13 @@ mod config {
         }
 
         #[test]
-        #[should_panic(expected = "InvalidPluginLocatorExt(\".wasm\")")]
+        #[should_panic(expected = "InvalidPluginLocatorExt(\".wasm OR .toml\")")]
         fn errors_for_invalid_url_ext() {
             PluginLocator::from_str("source:https://foo.com/file.yaml").unwrap();
         }
 
         #[test]
-        #[should_panic(expected = "InvalidPluginLocatorExt(\".wasm\")")]
+        #[should_panic(expected = "InvalidPluginLocatorExt(\".wasm OR .toml\")")]
         fn errors_for_invalid_path_ext() {
             PluginLocator::from_str("source:../foo/file.yaml").unwrap();
         }